# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NOOR AI Educational Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: videos
    description: Educational videos
  - name: chat
    description: AI Tutor chat
  - name: career
    description: Career roadmap
  - name: subscription
    description: Subscription management
  - name: progress
    description: Student progress
  - name: dashboard
    description: Dashboard summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List all videos
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: grade
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Add a new video (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoBody"
      responses:
        "201":
          description: Video created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"

  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a single video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateVideo
      tags: [videos]
      summary: Update a video (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoBody"
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /videos/{id}/checkpoints:
    get:
      operationId: getVideoCheckpoints
      tags: [videos]
      summary: Get AI checkpoints for a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checkpoints list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checkpoint"
    post:
      operationId: createCheckpoint
      tags: [videos]
      summary: Add an AI checkpoint (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckpointBody"
      responses:
        "201":
          description: Checkpoint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkpoint"

  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send message to AI Tutor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /career/roadmap:
    post:
      operationId: generateCareerRoadmap
      tags: [career]
      summary: Generate a career roadmap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerRoadmapBody"
      responses:
        "200":
          description: Career roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerRoadmap"

  /subscription/status:
    get:
      operationId: getSubscriptionStatus
      tags: [subscription]
      summary: Get current subscription status
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"

  /subscription/create:
    post:
      operationId: createSubscription
      tags: [subscription]
      summary: Create a new subscription (crypto/USDT via NOWPayments)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "200":
          description: Payment info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInfo"

  /progress:
    get:
      operationId: getProgress
      tags: [progress]
      summary: Get student progress
      responses:
        "200":
          description: Progress data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProgress"
    post:
      operationId: recordProgress
      tags: [progress]
      summary: Record video watch progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordProgressBody"
      responses:
        "200":
          description: Progress recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get student dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/admin:
    get:
      operationId: getAdminDashboard
      tags: [dashboard]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Admin dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        subscribed:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, role, subscribed, createdAt]

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required: [name, email, password]

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]

    Video:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        youtubeId:
          type: string
        subject:
          type: string
        grade:
          type: string
        duration:
          type: integer
        thumbnailUrl:
          type: string
        checkpointCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, youtubeId, subject, grade, duration, checkpointCount, createdAt]

    CreateVideoBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        youtubeId:
          type: string
        subject:
          type: string
        grade:
          type: string
        duration:
          type: integer
        thumbnailUrl:
          type: string
      required: [title, youtubeId, subject, grade, duration]

    Checkpoint:
      type: object
      properties:
        id:
          type: integer
        videoId:
          type: integer
        timestampSeconds:
          type: integer
        question:
          type: string
        context:
          type: string
      required: [id, videoId, timestampSeconds, question]

    CreateCheckpointBody:
      type: object
      properties:
        timestampSeconds:
          type: integer
        question:
          type: string
        context:
          type: string
      required: [timestampSeconds, question]

    ChatMessageBody:
      type: object
      properties:
        message:
          type: string
        videoId:
          type: integer
        checkpointId:
          type: integer
        history:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
            required: [role, content]
      required: [message]

    ChatResponse:
      type: object
      properties:
        reply:
          type: string
        understood:
          type: boolean
      required: [reply, understood]

    CareerRoadmapBody:
      type: object
      properties:
        interests:
          type: array
          items:
            type: string
        grades:
          type: object
          additionalProperties:
            type: string
        currentGrade:
          type: string
        goals:
          type: string
      required: [interests, currentGrade]

    CareerRoadmap:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapStep"
        estimatedYears:
          type: integer
        topCareers:
          type: array
          items:
            type: string
        recommendedSubjects:
          type: array
          items:
            type: string
      required: [title, summary, steps, estimatedYears, topCareers, recommendedSubjects]

    RoadmapStep:
      type: object
      properties:
        phase:
          type: integer
        title:
          type: string
        description:
          type: string
        duration:
          type: string
        milestones:
          type: array
          items:
            type: string
      required: [phase, title, description, duration, milestones]

    SubscriptionStatus:
      type: object
      properties:
        subscribed:
          type: boolean
        plan:
          type: string
        expiresAt:
          type: string
        daysRemaining:
          type: integer
      required: [subscribed]

    CreateSubscriptionBody:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [crypto, card]
        currency:
          type: string
      required: [paymentMethod]

    PaymentInfo:
      type: object
      properties:
        paymentId:
          type: string
        paymentUrl:
          type: string
        paymentAddress:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        expiresAt:
          type: string
      required: [paymentId, amount, currency, status]

    RecordProgressBody:
      type: object
      properties:
        videoId:
          type: integer
        watchedSeconds:
          type: integer
        completed:
          type: boolean
      required: [videoId, watchedSeconds]

    StudentProgress:
      type: object
      properties:
        totalVideosWatched:
          type: integer
        totalMinutesLearned:
          type: integer
        completedVideos:
          type: array
          items:
            type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ProgressEntry"
      required: [totalVideosWatched, totalMinutesLearned, completedVideos, recentActivity]

    ProgressEntry:
      type: object
      properties:
        videoId:
          type: integer
        videoTitle:
          type: string
        watchedSeconds:
          type: integer
        completed:
          type: boolean
        updatedAt:
          type: string
      required: [videoId, videoTitle, watchedSeconds, completed, updatedAt]

    DashboardSummary:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        totalVideos:
          type: integer
        completedVideos:
          type: integer
        minutesLearned:
          type: integer
        streakDays:
          type: integer
        recentVideos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        subscribed:
          type: boolean
      required: [totalVideos, completedVideos, minutesLearned, streakDays, recentVideos, subscribed]

    AdminDashboard:
      type: object
      properties:
        totalUsers:
          type: integer
        totalVideos:
          type: integer
        totalSubscribers:
          type: integer
        recentUsers:
          type: array
          items:
            $ref: "#/components/schemas/User"
      required: [totalUsers, totalVideos, totalSubscribers, recentUsers]