# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NOOR AI Educational Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";
import type { Video } from "./video";

export interface DashboardSummary {
  user?: User;
  totalVideos: number;
  completedVideos: number;
  minutesLearned: number;
  streakDays: number;
  recentVideos: Video[];
  subscribed: boolean;
}