# lib/api-zod/src/generated/types/paymentInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NOOR AI Educational Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInfo {
  paymentId: string;
  paymentUrl?: string;
  paymentAddress?: string;
  amount: number;
  currency: string;
  status: string;
  expiresAt?: string;
}