# lib/api-zod/src/generated/types/progressEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NOOR AI Educational Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressEntry {
  videoId: number;
  videoTitle: string;
  watchedSeconds: number;
  completed: boolean;
  updatedAt: string;
}